refactor: extract generatePdf service
