+++ tests/ucmoGeminiLayout.test.js
@@ -8,7 +8,7 @@ const launch = jest.fn().mockResolvedValue({ newPage, close });
 
 jest.unstable_mockModule('puppeteer', () => ({ default: { launch } }));
 
-const { generatePdf } = await import('../server.js');
+const { generatePdf } = await import('../services/generatePdf.js');
 
 test('ucmo template uses Gemini layout markup', async () => {
   const markup = '<html><body><table id="top"><tr><td>contact</td><td><img src="logo.png"/></td></tr></table></body></html>';
